requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.18.0
datasets>=2.0.0
tqdm>=4.62.0
pillow>=8.3.0
requests>=2.26.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0
sacremoses
rouge